代码结构优化
