+++ app/src/main/java/com/cofbro/qian/mapsetting/util/PreWeb.kt
@@ -1,0 +1,10 @@
+package com.cofbro.qian.mapsetting.util
+
+import org.jsoup.nodes.Document
+
+class PreWeb(
+    val html: Document,
+    val locationText:String?,
+    val latitude:String?,
+    val longitude:String?,
+    val statusContent:String)
